backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo==4.6.1
pymilvus==2.3.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
bcrypt==4.0.1
openai==1.3.0
langchain==0.0.335
tiktoken==0.5.1
python-magic==0.4.27
httpx==0.25.0
email-validator==2.0.0
jieba>=0.42.1
PyPDF2>=3.0.0
redis>=4.0.0
numpy>=1.21.0
pytest>=7.0.0
pytest-cov>=4.0.0
psutil>=5.9.0 